Changed amount type to double

+++ library/src/main/java/com/step/bank/Transaction.java
@@ -4,11 +4,11 @@ import java.util.Date;
 import java.util.Objects;
 
 public class Transaction {
-  private final float amount;
+  private final double amount;
   private final String to;
   private Date date;
 
-  Transaction(float amount, Date date, String to) {
+  Transaction(double amount, Date date, String to) {
     this.amount = amount;
     this.date = date;
     this.to = to;
@@ -23,7 +23,7 @@ public class Transaction {
     if (this == o) return true;
     if (o == null || getClass() != o.getClass()) return false;
     Transaction that = (Transaction) o;
-    return Float.compare(that.amount, amount) == 0 &&
+    return Double.compare(that.amount, amount) == 0 &&
         Objects.equals(to, that.to) &&
         Objects.equals(date, that.date);
   }
