+++ library/src/main/java/com/step/bank/Transactions.java
@@ -10,7 +10,7 @@ public class Transactions {
     this.list = new ArrayList<>();
   }
 
-  public void debit(float amount, String name) {
+  public void debit(double amount, String name) {
     this.list.add(new DebitTransaction(amount, name));
   }
 }
